requirements.txt
alembic==1.7.5
aniso8601==9.0.1
astroid==2.9.3
click==8.0.3
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mccabe==0.6.1
numpy==1.22.1
numpy-financial==1.0.0
packaging==21.3
platformdirs==2.4.1
psycopg2==2.9.3
pylint==2.12.2
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
typing_extensions==4.0.1
webargs==8.1.0
Werkzeug==2.0.2
wrapt==1.13.3
yapf==0.32.0
